feat: move useProjects hook to a dedicated file for better organization
